+++ components/menu/Menu.js
@@ -262,9 +262,9 @@ class Menu extends Component {
   }
 
   render () {
-    const {data, placement, showCollapse} = this.props
+    const {data, placement, showCollapse, theme} = this.props
     const {collapsed} = this.state
-    const cls = classNames('hi-menu', `hi-menu--${placement}`, {
+    const cls = classNames('hi-menu', `theme__${theme}`, `hi-menu--${placement}`, {
       'hi-menu--mini': collapsed
     })
     const miniIcon = <i className={`hi-icon icon-${collapsed ? 'Expand' : 'Collapse'}`} />