+++ components/upload/UploadDrag.js
@@ -46,7 +46,8 @@ class UploadDrag extends Upload {
       disabled,
       tips,
       localeDatas,
-      onRemove
+      onRemove,
+      theme
     } = this.props
     const {
       overEvent,
@@ -55,6 +56,7 @@ class UploadDrag extends Upload {
     } = this.state
 
     const dragCls = classNames(
+      `theme__${theme}`,
       'hi-upload',
       'hi-upload--drag',
       overEvent && !disabled && 'drop-over',