+++ docs/demo/button/section-link.jsx
@@ -2,7 +2,11 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Button from '../../../components/button'
 const prefix = 'button-link'
-const rightOptions = ['正常', '禁用']
+const rightOptions = ['文本链接', '带图标链接']
+const desc = [
+  '文本链接：执行操作时发出页面请求，页面会给予明确的反馈',
+  '带图标链接：强调动作含义或简化动作的操作区域'
+]
 const code = [
   {
     code: `import React from 'react'
@@ -11,18 +15,15 @@ class Demo extends React.Component {
   render () {
     return (
       <React.Fragment>
-        <Button type="primary" appearance="link">突出链接</Button>
-        <Button type="default" appearance="link">默认链接</Button>
-        <Button type="success" appearance="link">成功链接</Button>
-        <Button type="danger" appearance="link">危险链接</Button>
-        <Button type="line" appearance="link" icon="edit" />
-        <Button type="line" appearance="link" icon="down">展开更多</Button>
-        <Button type="line" appearance="link" icon="plus">新增筛选项</Button>
+        <Button type="primary" appearance="link">编辑</Button>
+        <Button type="default" appearance="link">查看更多</Button>
+        <Button type="success" appearance="link">复制链接到微信</Button>
+        <Button type="danger" appearance="link">删除</Button>
       </React.Fragment>
     )
   }
 }`,
-    opt: ['正常']
+    opt: ['文本链接']
   },
   {
     code: `import React from 'react'
@@ -31,18 +32,14 @@ class Demo extends React.Component {
   render () {
     return (
       <React.Fragment>
-        <Button disabled type="primary" appearance="link">突出链接</Button>
-        <Button disabled type="default" appearance="link">默认链接</Button>
-        <Button disabled type="success" appearance="link">成功链接</Button>
-        <Button disabled type="danger" appearance="link">危险链接</Button>
-        <Button disabled type="line" appearance="link" icon="edit" />
-        <Button disabled type="line" appearance="link" icon="down">展开更多</Button>
-        <Button disabled type="line" appearance="link" icon="plus">新增筛选项</Button>
+        <Button type="line" appearance="link" icon="edit" />
+        <Button type="line" appearance="link" icon="down">展开更多</Button>
+        <Button type="line" appearance="link" icon="plus">新增筛选项</Button>
       </React.Fragment>
     )
   }
 }`,
-    opt: ['禁用']
+    opt: ['带图标链接']
   }
 ]
 
@@ -52,6 +49,7 @@ const DemoLink = () => (
     scope={{ Button }}
     prefix={prefix}
     rightOptions={rightOptions}
+    desc={desc}
   />
 )
 export default DemoLink