+++ docs/demo/dropdown/section-type.jsx
@@ -2,8 +2,11 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import { Dropdown, Message } from '../../../components'
 const prefix = 'dropdown-type'
-const rightOptions = ['文字型', '按钮型', '按钮组']
-const desc = '通过 type 属性切换按钮类型'
+const rightOptions = ['按钮样式', '按钮动作']
+const desc = [
+  '按钮样式：以按钮为下拉菜单的入口样式，强调动作引导',
+  '按钮动作：一级动作前置曝光，减少操作成本'
+]
 
 const code = [
   {
@@ -75,7 +78,7 @@ class Demo extends React.Component {
     )
   }
 }`,
-    opt: ['按钮型']
+    opt: ['按钮样式']
   },
   {
     code: `import React from 'react'
@@ -114,7 +117,7 @@ class Demo extends React.Component {
     )
   }
 }`,
-    opt: ['按钮组']
+    opt: ['按钮动作']
   }
 ]