+++ docs/demo/form/section-two-col.jsx
@@ -5,8 +5,14 @@ import Input from '../../../components/input'
 import Button from '../../../components/button'
 import Grid from '../../../components/grid'
 const prefix = 'form-two-col'
-const desc = '当表单项数量较多、界面空间充裕且label名称较长时使用，两列间距可灵活调整。'
-const code = `import React from 'react'
+const rightOptions = ['上下结构', '左右结构']
+const desc = [
+  '上下结构：表单项对应的标题长度较长，表单项较多',
+  '左右结构：屏幕分辨率较大，表单项较多且对应标题长度短且易对齐'
+]
+const code = [
+  {
+    code: `import React from 'react'
 import { Form, Input, Grid } from '@hi-ui/hiui'
 class Demo extends React.Component {
   render (){
@@ -15,32 +21,32 @@ class Demo extends React.Component {
     const FormItem = Form.Item
     return (
       <div style={{width: 500}}>
-        <Form labelPlacement='top'>
+        <Form labelWidth='70' labelPlacement='top'>
           <Row>
             <Col span={12}>
-              <FormItem label='姓名' labelWidth='70'>
+              <FormItem label='姓名' >
                 <Input placeholder={'请输入'} style={{width: '200'}} />
               </FormItem>
             </Col>
             <Col span={12}>
-              <FormItem label='密码' labelWidth='70'>
+              <FormItem label='密码' >
                 <Input type='password' placeholder={'请输入'} style={{width: '200'}} />
               </FormItem>
             </Col>
           </Row>
           <Row>
             <Col span={12}>
-              <FormItem label='手机号' labelWidth='70'>
+              <FormItem label='手机号' >
                 <Input placeholder={'请输入'} style={{width: '200'}} />
               </FormItem>
             </Col>
             <Col span={12}>
-              <FormItem label='邮箱' labelWidth='70'>
+              <FormItem label='邮箱' >
                 <Input type='password' placeholder={'请输入'} style={{width: '200'}} />
               </FormItem>
             </Col>
           </Row>
-          <FormItem>
+          <FormItem >
             <Button type='primary' onClick={()=>{}}>提交</Button>
             <Button type='line' onClick={()=>{}}>重置</Button>
           </FormItem>
@@ -48,7 +54,56 @@ class Demo extends React.Component {
       </div>
     )
   }
-}`
+}`,
+    opt: ['上下结构']
+  },
+  {
+    code: `import React from 'react'
+import { Form, Input, Grid } from '@hi-ui/hiui'
+class Demo extends React.Component {
+  render (){
+    const Row = Grid.Row
+    const Col = Grid.Col
+    const FormItem = Form.Item
+    return (
+      <div style={{width: 500}}>
+        <Form labelWidth='70' labelPlacement='left'>
+          <Row>
+            <Col span={12}>
+              <FormItem label='姓名' >
+                <Input placeholder={'请输入'} style={{width: '200'}} />
+              </FormItem>
+            </Col>
+            <Col span={12}>
+              <FormItem label='密码' >
+                <Input type='password' placeholder={'请输入'} style={{width: '200'}} />
+              </FormItem>
+            </Col>
+          </Row>
+          <Row>
+            <Col span={12}>
+              <FormItem label='手机号' >
+                <Input placeholder={'请输入'} style={{width: '200'}} />
+              </FormItem>
+            </Col>
+            <Col span={12}>
+              <FormItem label='邮箱' >
+                <Input type='password' placeholder={'请输入'} style={{width: '200'}} />
+              </FormItem>
+            </Col>
+          </Row>
+          <FormItem >
+            <Button type='primary' onClick={()=>{}}>提交</Button>
+            <Button type='line' onClick={()=>{}}>重置</Button>
+          </FormItem>
+        </Form>
+      </div>
+    )
+  }
+}`,
+    opt: ['左右结构']
+  }
+]
 
 const DemoRow = () => (
   <DocViewer
@@ -56,6 +111,7 @@ const DemoRow = () => (
     scope={{ Form, Button, Input, Grid }}
     prefix={prefix}
     desc={desc}
+    rightOptions={rightOptions}
   />
 )
 export default DemoRow