+++ libs/doc-viewer/components/EditorWrapper.jsx
@@ -11,7 +11,9 @@ export default class EditorWrapper extends React.Component {
     descBarHeight: 40
   }
   componentDidMount () {
-    const descBar = document.getElementsByClassName(`${this.props.prefix}-desc-bar`)[0]
+    const descBar = document.getElementsByClassName(
+      `${this.props.prefix}-desc-bar`
+    )[0]
     this.setState({ descBarHeight: descBar.clientHeight })
     this.setInnerHeight()
     const clipboard = new Clipboard(`.${this.props.prefix}-copy-btn`)
@@ -29,7 +31,9 @@ export default class EditorWrapper extends React.Component {
     this.setInnerHeight()
   }
   setInnerHeight = () => {
-    const codeViewer = document.getElementsByClassName(`${this.props.prefix}-editor-inner`)[0]
+    const codeViewer = document.getElementsByClassName(
+      `${this.props.prefix}-editor-inner`
+    )[0]
     this.setState({
       innerHeight: codeViewer.clientHeight
     })
@@ -58,7 +62,9 @@ export default class EditorWrapper extends React.Component {
       <div
         className='editor-wrapper'
         style={{
-          height: this.state.collapse ? innerHeight + descBarHeight : descBarHeight,
+          height: this.state.collapse
+            ? innerHeight + descBarHeight
+            : descBarHeight,
           overflow: 'hidden'
         }}
       >
@@ -68,13 +74,14 @@ export default class EditorWrapper extends React.Component {
             minHeight: 40,
             display: 'flex',
             justifyContent: 'space-between',
-            paddingLeft: 20,
-            paddingRight: 20,
-            alignItems: 'center',
+            padding: '10px 20px',
+            boxSizing: 'border-box',
             borderBottom: this.state.collapse ? '1px dashed #e6e7e8' : 'none'
           }}
         >
-          <div style={{ maxWidth: 720, paddingTop: 10, paddingBottom: 10 }}>{desc}</div>
+          <div style={{ flex: 1 }}>
+            {Array.isArray(desc) ? desc.map(d => <div>{d}</div>) : desc}
+          </div>
           <div>
             <Tooltip
               title={this.state.collapse ? '收起代码' : '展开代码'}
@@ -87,17 +94,27 @@ export default class EditorWrapper extends React.Component {
                   })
                 }}
               >
-                {this.state.collapse ? <Icon name='api' /> : <Icon name='noapi' />}
+                {this.state.collapse ? (
+                  <Icon name='api' />
+                ) : (
+                  <Icon name='noapi' />
+                )}
               </span>
             </Tooltip>
             {copyed ? (
-              <Tooltip title='复制成功' style={{ margin: '0 8px', cursor: 'pointer' }}>
+              <Tooltip
+                title='复制成功'
+                style={{ margin: '0 8px', cursor: 'pointer' }}
+              >
                 <span>
                   <Icon name='check' />
                 </span>
               </Tooltip>
             ) : (
-              <Tooltip title='复制代码' style={{ margin: '0 8px', cursor: 'pointer' }}>
+              <Tooltip
+                title='复制代码'
+                style={{ margin: '0 8px', cursor: 'pointer' }}
+              >
                 <span
                   className={`${this.props.prefix}-copy-btn`}
                   data-clipboard-target={`.${prefix}-editor-inner .npm__react-simple-code-editor__textarea`}
@@ -106,7 +123,10 @@ export default class EditorWrapper extends React.Component {
                 </span>
               </Tooltip>
             )}
-            <Tooltip title='重置代码' style={{ margin: '0 8px', cursor: 'pointer' }}>
+            <Tooltip
+              title='重置代码'
+              style={{ margin: '0 8px', cursor: 'pointer' }}
+            >
               <span
                 onClick={() => {
                   this.editor.updateContent(code)